requirements.txt
digitalai-release-sdk==25.1.1
fastmcp==2.12.4
google-genai==1.41.0
langchain-core==0.3.78
langchain-mcp-adapters==0.1.11
langgraph==0.6.8
langchain-google-genai==2.1.12